fix position modal
