--- src/pages/DetailProduct.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function DetailProduct() {
-  return <h1>Detalles del producto</h1>;
-}
-
-export default DetailProduct;
