--- src/pages/DetailProductContainer.js
@@ -1,12 +1,0 @@
-import React from "react";
-import { Outlet } from "react-router-dom";
-
-function DetailProductContainer() {
-  return (
-    <div>
-      <Outlet />
-    </div>
-  );
-}
-
-export default DetailProductContainer;
