--- src/pages/NewProduct.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function NewProduct() {
-  return <h1>New Product</h1>;
-}
-
-export default NewProduct;
