--- src/pages/UpdateProduct.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function UpdateProduct() {
-  return <h1>Update Product</h1>;
-}
-
-export default UpdateProduct;
